Add support for new FLIM LABS JSON format
