+++ src/phasorpy/datasets.py
@@ -310,53 +310,61 @@ FLIMLABS = pooch.create(
     base_url='https://github.com/phasorpy/phasorpy-data/raw/main/flimlabs',
     env=ENV,
     registry={
-        'calibrator2_imaging.json': (
+        'calibrator_2_5_1737112045_imaging.json': (
             'sha256:'
-            '7bd575a9ed0cc6e4b5cd4ac236a6414443d496420d3fbc2c0474e95d9b65f046'
+            'a34c7077e88d1e7272953a46b2bb4e3ab8adf5a2f61c824dfc27032d952b920e'
         ),
-        'calibrator2_imaging.json.zip': (
+        'calibrator_2_5_1737112045_imaging.json.zip': (
             'sha256:'
-            '7423d1faa9644e8e0dc1da5e3ce49811383487e0c17cfccf1030c250ecc6dada'
+            'fea791b28afd8365152018810cbbaaac1177cb72827578073587a1050d1af329'
         ),
-        'calibrator2_imaging_calibration.json': (
+        'calibrator_2_5_1737112045_imaging_calibration.json': (
             'sha256:'
-            '0d0481d1767336fa5e5dce8584b30bc83ac5d84e94514b87e4aa491f55acf59b'
+            '8f2ebe9b544fae9524dc13221c1a5ab1b57d9dfd40ec2eb06a7b1475fcd63057'
         ),
-        'calibrator2_imaging_calibration.json.zip': (
+        'calibrator_2_5_bis_1737112494_imaging.json': (
             'sha256:'
-            '77235ba78202ebc774f4ee20f0ffb496461596191f7bee5bc5c9a8d1c54e9da3'
+            '0509c5aba066419b03f83264eba58acbf4aae470aa1057c52f45e60225e033a4'
         ),
-        'dataset_1.json': (
+        'calibrator_2_5_bis_1737112494_imaging.json.zip': (
             'sha256:'
-            'fe7c1d06d557d0bf5c002e3d4950b66e390d758d6517cb4d6c7fbd0525fc0635'
+            'bdc5df2a3f08a64ec7b7bb57b36e21546de142e67c59c052318252dbb66d8abf'
         ),
-        'dataset_1.json.zip': (
+        'calibrator_2_5_bis_1737112494_imaging_calibration.json': (
             'sha256:'
-            '0ea4dd9204a6f7d08f41b0f4c53005a452cb3d1a9d3291a805d2ff0433d83e3e'
+            '9bb0e21b1e7c04add672aa8a78048b09908c860fcaf907ca33c0c87d161f6ebf'
         ),
-        'dataset_1_phasor_ch1_h1.json': (
+        'convallaria_1_1737112980_phasor_ch1.json': (
             'sha256:'
-            '56f79d582e0af58aa256a59ab55b85395d865bcd7b126b78a54da63079ebe3a0'
+            '4a296a0d7898dc660a388e1bba5cf98b43c35fe12d94b7aba48d00245e37242d'
         ),
-        'dataset_1_phasor_ch1_h1.json.zip': (
+        'convallaria_1_1737112980_phasor_ch1.json.zip': (
             'sha256:'
-            '8cfaf663e0716d5f8c612d213acc554e7e863b33d6dc1519683abe9f020e188b'
+            '79c416b9099c9f58d2092fe5b26ea6d0f695977b877784cf564d3ead896d9354'
         ),
-        'dataset_2.json': (
+        'convallaria_2_1737113097_phasor_ch1.json': (
             'sha256:'
-            '0b1f3034bc46b0f74cbf3c9a9a231ff877b3372cd3806aa5c2f7d443f258b448'
+            'da549645ffd898238c26f7a1eac3aca4ffccec86653c0d241a6ece674dfce90d'
         ),
-        'dataset_2.json.zip': (
+        'convallaria_2_1737113097_phasor_ch1.json.zip': (
             'sha256:'
-            '57e9fdf456efbe92327ced7d71a009302b08ca459226c69877a2857453b5dfeb'
+            '8801bb14b457dceaef42e8b3bf6af770a2e14264cd2b282ba7e3d70b91ea954c'
         ),
-        'dataset_2_phasor_ch1_h1.json': (
+        'data_2_calibrator_2_5_1737112409_phasor_ch1.json': (
             'sha256:'
-            '9bd26dfe7cbb62befc9d6bd86336d698287a08c0fe2677c504876c7a24eba55a'
+            'ea8683892eb76f52231e5d6ceab64a3737454aa95fe73185366de8f758fd9b70'
         ),
-        'dataset_2_phasor_ch1_h1.json.zip': (
+        'data_2_calibrator_2_5_1737112409_phasor_ch1.json.zip': (
             'sha256:'
-            '427dcffe43ef99f1e2b7d750af822b15b3414fb8da5afa6efb964228751fa180'
+            '40d2aa90b95fd8864a2392337c83a1a4f4931d7359cb30a486f65173f208de0a'
+        ),
+        'data_calibrator_2_5_1737112133_phasor_ch1.json': (
+            'sha256:'
+            '6a8790212bc62014d597402ec5feb0e50ec6ae2aa62d63fae8cb62c6c5656268'
+        ),
+        'data_calibrator_2_5_1737112133_phasor_ch1.json.zip': (
+            'sha256:'
+            'c9ef343bdbd7a51d23fdf4082e379dcdb1ce9f3e2ba065289bf2925d68ef55ba'
         ),
     },
 )
